+++ spec/rubocop/cop/corrector_spec.rb
@@ -1,6 +1,29 @@
 # frozen_string_literal: true
 
 describe RuboCop::Cop::Corrector do
+  { replace: 'hello',
+    remove: nil,
+    insert_before: 'hi',
+    insert_after: '!',
+  }.each do |method, arg|
+    describe "##{method}" do
+      it 'requires ranges with the same source_buffer' do
+        processed_source = parse_source('2 * 21')
+        corrector = described_class.new(processed_source.buffer)
+        other_source = parse_source('42')
+
+        expect {
+          corrector.send(method, other_source.ast.loc.expression, *arg)
+        }.to raise_error(ArgumentError)
+
+        # Sanity check that test works with same buffer:
+        expect {
+          corrector.send(method, processed_source.ast.loc.expression, *arg)
+        }.not_to raise_error
+      end
+    end
+  end
+
   describe '#rewrite' do
     it 'allows removal of a range' do
       source = 'true and false'